llm/requirements.txt
python-dotenv
# gpt_index==0.4.24
langchain==0.0.148
# llama-index==0.5.6
gradio
llama-index
ipython
# langchain